Complete Phase 2: items 2.1 (SessionService constructor), 2.2 (split command.ts)
2.1: Add typed GovernanceEngine constructor param to SessionService.
     Remove (sessionService as any).governanceEngine mutation from app.ts.
2.2: Split 836-line command.ts into 11 focused files under handlers/commands/.
     command.ts reduced to thin dispatcher (~113 lines).
