+++ packages/slack-bot/src/handlers/commands/metadata.ts
@@ -1,0 +1,2 @@
+// Placeholder for future !capture-type and !brain-view command group handlers.
+// Add handleCaptureType and handleBrainView functions here when these commands are implemented.
