+++ packages/slack-bot/src/handlers/commands/search.ts
@@ -1,0 +1,3 @@
+// Placeholder for future !search command group handlers.
+// Search is currently handled via synthesis.ts and query.ts routes.
+// Add handleSearch and related functions here when !search commands are implemented.
