+++ packages/slack-bot/src/handlers/commands/skill.ts
@@ -1,0 +1,2 @@
+// Placeholder for future !skill command group handlers.
+// Add handleSkillList, handleSkillTrigger, etc. here when !skill commands are implemented.
